+++ atomicshop/python_functions.py
@@ -1,4 +1,3 @@
-# v1.0.1 - 21.03.2023 13:40
 import sys
 
 from .print_api import print_api